+++ programs/clearing_house/src/validation/order.rs
@@ -14,7 +14,7 @@ use crate::validate;
 pub fn validate_order(
     order: &Order,
     market: &PerpMarket,
-    valid_oracle_price: Option<i128>,
+    valid_oracle_price: Option<i64>,
     slot: u64,
 ) -> ClearingHouseResult {
     match order.order_type {
@@ -90,7 +90,7 @@ fn validate_market_order(
 fn validate_limit_order(
     order: &Order,
     market: &PerpMarket,
-    valid_oracle_price: Option<i128>,
+    valid_oracle_price: Option<i64>,
     slot: u64,
 ) -> ClearingHouseResult {
     validate_base_asset_amount(
@@ -138,7 +138,7 @@ fn validate_limit_order(
 fn validate_post_only_order(
     order: &Order,
     market: &PerpMarket,
-    valid_oracle_price: Option<i128>,
+    valid_oracle_price: Option<i64>,
     slot: u64,
 ) -> ClearingHouseResult {
     let base_asset_amount_market_can_fill = calculate_base_asset_amount_to_fill_up_to_limit_price(
@@ -253,7 +253,7 @@ fn validate_base_asset_amount(
 
 pub fn validate_spot_order(
     order: &Order,
-    valid_oracle_price: Option<i128>,
+    valid_oracle_price: Option<i64>,
     slot: u64,
     step_size: u64,
     tick_size: u64,
@@ -284,7 +284,7 @@ pub fn validate_spot_order(
 
 fn validate_spot_limit_order(
     order: &Order,
-    valid_oracle_price: Option<i128>,
+    valid_oracle_price: Option<i64>,
     slot: u64,
     step_size: u64,
     tick_size: u64,
